add percepts subpackage
