+++ examples/models/plot_scoreboard.py
@@ -141,7 +141,7 @@ percept = model.predict_percept(implant)
 
 ##############################################################################
 # The resulting percept is stored in a
-# :py:class:`~pulse2percept.models.Percept` object, which is similar in
+# :py:class:`~pulse2percept.percepts.Percept` object, which is similar in
 # organization to the :py:class:`~pulse2percept.stimuli.Stimulus` object:
 # the ``data`` container is a 3D NumPy array (Y, X, T) with labeled axes
 # ``xdva``, ``ydva``, and ``time``.
@@ -152,11 +152,11 @@ ax = percept.plot()
 ax.set_title('Predicted percept')
 
 ##############################################################################
-# By default, the :py:meth:`~pulse2percept.models.Percept.plot` method uses
+# By default, the :py:meth:`~pulse2percept.percepts.Percept.plot` method uses
 # Matplotlib's ``pcolor`` function.
 # However, it can also be configured to use a hex grid (using Matplotlib's
 # ``hexbin`` function). Additional parameters can be passed to ``hexbin`` as
-# keyword arguments of :py:meth:`~pulse2percept.models.Percept.plot`:
+# keyword arguments of :py:meth:`~pulse2percept.percepts.Percept.plot`:
 
 implant.stim = np.arange(60)
 percept = model.predict_percept(implant)
