+++ pulse2percept/percepts/__init__.py
@@ -1,0 +1,13 @@
+"""Visual percepts and phosphenes.
+
+.. autosummary::
+    :toctree: _api
+
+    base
+
+"""
+from .base import Percept
+
+__all__ = [
+    'Percept'
+]
