+++ pulse2percept/percepts/tests/test_base.py
@@ -1,0 +1,87 @@
+import numpy as np
+import pytest
+import numpy.testing as npt
+import matplotlib.pyplot as plt
+from matplotlib.axes import Subplot
+
+from pulse2percept.percepts import Percept
+from pulse2percept.utils import GridXY
+
+
+def test_Percept():
+    # Automatic axes:
+    ndarray = np.arange(15).reshape((3, 5, 1))
+    percept = Percept(ndarray, metadata='meta')
+    npt.assert_equal(percept.shape, ndarray.shape)
+    npt.assert_equal(percept.metadata, 'meta')
+    npt.assert_equal(hasattr(percept, 'xdva'), True)
+    npt.assert_almost_equal(percept.xdva, np.arange(ndarray.shape[1]))
+    npt.assert_equal(hasattr(percept, 'ydva'), True)
+    npt.assert_almost_equal(percept.ydva, np.arange(ndarray.shape[0]))
+    npt.assert_equal(hasattr(percept, 'time'), True)
+    npt.assert_almost_equal(percept.time, np.arange(ndarray.shape[2]))
+
+    # Specific labels:
+    percept = Percept(ndarray, time=0.4)
+    npt.assert_almost_equal(percept.time, [0.4])
+    percept = Percept(ndarray, time=[0.4])
+    npt.assert_almost_equal(percept.time, [0.4])
+
+    # Labels from a grid.
+    y_range = (-1, 1)
+    x_range = (-2, 2)
+    grid = GridXY(x_range, y_range)
+    percept = Percept(ndarray, space=grid)
+    npt.assert_almost_equal(percept.xdva, grid._xflat)
+    npt.assert_almost_equal(percept.ydva, grid._yflat)
+    npt.assert_almost_equal(percept.time, [0])
+    grid = GridXY(x_range, y_range)
+    percept = Percept(ndarray, space=grid)
+    npt.assert_almost_equal(percept.xdva, grid._xflat)
+    npt.assert_almost_equal(percept.ydva, grid._yflat)
+    npt.assert_almost_equal(percept.time, [0])
+
+
+def test_Percept_plot():
+    y_range = (-1, 1)
+    x_range = (-2, 2)
+    grid = GridXY(x_range, y_range)
+    percept = Percept(np.arange(15).reshape((3, 5, 1)), space=grid)
+
+    # Basic usage of pcolor:
+    ax = percept.plot(kind='pcolor')
+    npt.assert_equal(isinstance(ax, Subplot), True)
+    npt.assert_almost_equal(ax.axis(), [0, len(percept.xdva),
+                                        0, len(percept.ydva)])
+    npt.assert_almost_equal(ax.collections[0].get_clim(),
+                            [percept.data.min(), percept.data.max()])
+
+    # Basic usage of hex:
+    ax = percept.plot(kind='hex')
+    npt.assert_equal(isinstance(ax, Subplot), True)
+    npt.assert_almost_equal(ax.axis(), [percept.xdva[0], percept.xdva[-1],
+                                        percept.ydva[0], percept.ydva[-1]])
+    npt.assert_almost_equal(ax.collections[0].get_clim(),
+                            [percept.data[..., 0].min(),
+                             percept.data[..., 0].max()])
+
+    # Verify color map:
+    npt.assert_equal(ax.collections[0].cmap, plt.cm.gray)
+    ax = percept.plot(cmap='inferno')
+    npt.assert_equal(ax.collections[0].cmap, plt.cm.inferno)
+
+    # Invalid calls:
+    with pytest.raises(ValueError):
+        percept.plot(kind='invalid')
+    with pytest.raises(TypeError):
+        percept.plot(ax='invalid')
+
+    # TODO
+    with pytest.raises(NotImplementedError):
+        percept.plot(time=3.3)
+
+
+def test_Percept_get_brightest_frame():
+    percept = Percept(np.arange(30).reshape((3, 5, 2)))
+    npt.assert_almost_equal(percept.get_brightest_frame(),
+                            percept.data[..., 1])
